requirements.txt
fastapi
uvicorn
kafka-python
pydantic
